requirements.txt
Flask==1.1.1
gunicorn==19.9.0
line-bot-sdk==1.14.0
psycopg2==2.8.3
mysql-connector==2.2.9
